+++ src/UI/Dialogs/GradientDialog.gd
@@ -9,6 +9,7 @@ onready var preview : TextureRect = $VBoxContainer/Preview
 onready var color1 : ColorPickerButton = $VBoxContainer/ColorsContainer/ColorPickerButton
 onready var color2 : ColorPickerButton = $VBoxContainer/ColorsContainer/ColorPickerButton2
 onready var steps : SpinBox = $VBoxContainer/StepsContainer/StepSpinBox
+onready var direction : OptionButton = $VBoxContainer/DirectionContainer/DirectionOptionButton
 
 
 func _on_GradientDialog_about_to_show() -> void:
@@ -21,7 +22,7 @@ func _on_GradientDialog_about_to_show() -> void:
 
 func update_preview() -> void:
 	preview_image.copy_from(current_cel)
-	DrawingAlgos.generate_gradient(preview_image, [color1.color, color2.color], steps.value)
+	DrawingAlgos.generate_gradient(preview_image, [color1.color, color2.color], steps.value, direction.selected)
 	preview_texture.create_from_image(preview_image, 0)
 	preview.texture = preview_texture
 
@@ -38,8 +39,12 @@ func _on_StepSpinBox_value_changed(_value : int) -> void:
 	update_preview()
 
 
+func _on_OptionButton_item_selected(_index : int) -> void:
+	update_preview()
+
+
 func _on_GradientDialog_confirmed() -> void:
 	Global.canvas.handle_undo("Draw")
-	DrawingAlgos.generate_gradient(current_cel, [color1.color, color2.color], steps.value)
+	DrawingAlgos.generate_gradient(current_cel, [color1.color, color2.color], steps.value, direction.selected)
 	Global.canvas.update_texture(Global.current_project.current_layer)
 	Global.canvas.handle_redo("Draw")